Add styles for HomePage and FirstCard components; refactor Header layout and update WeatherData interface
