+++ src/components/FirstCard/FirstCard.scss
@@ -1,0 +1,2 @@
+@use "../../assets/styles/variables" as vF;
+@use "../../assets/styles/mixins" as mF;
